Integrate LDT oracles into main transrcipt (#7)
This PR refactors some of inner logic of LDT. Basically, now, LDT will not have a separate transcript. Instead, LDT will keep track of `MsgRoundRef` of each codeword oracle, and work on the same transcript as user's code. This can greatly improve readability because we do not need to interact with oracles outside the transcript. 

No breaking change in APIs for user.